refactor(batcher): rename files
