+++ crates/batcher/src/lib.rs
@@ -2,6 +2,6 @@ pub mod batcher;
 pub mod communication;
 pub mod config;
 pub mod fee_market;
-pub mod proposals_manager;
+pub mod proposal_manager;
 #[cfg(test)]
-mod proposals_manager_test;
+mod proposal_manager_test;
