+++ crates/batcher/src/proposal_manager.rs
